refactor(feature-flags): move getAllFeatureFlags to data/utils
